+++ include/barry/barraydense-meat-operators.hpp
@@ -38,8 +38,8 @@ BDENSE_TEMPLATE(BDENSE_TYPE()&, operator+=) (
     // Must be compatible
     checkdim_(*this, rhs);
     
-    for (uint i = 0u; i < nrow(); ++i)
-        for (uint j = 0u; j < ncol(); ++j)
+    for (size_t i = 0u; i < nrow(); ++i)
+        for (size_t j = 0u; j < ncol(); ++j)
             this->operator()(i, j) += rhs.get_cell(i, j);
 
     return *this;
@@ -49,8 +49,8 @@ BDENSE_TEMPLATE(BDENSE_TYPE()&, operator+=) (
     const Cell_Type& rhs
 ) {
 
-    for (uint i = 0u; i < nrow(); ++i) {
-        for (uint j = 0u; j < ncol(); ++j) {
+    for (size_t i = 0u; i < nrow(); ++i) {
+        for (size_t j = 0u; j < ncol(); ++j) {
             this->operator()(i, j) += rhs;
         }
     }
@@ -65,8 +65,8 @@ BDENSE_TEMPLATE(BDENSE_TYPE()&, operator-=) (
     // Must be compatible
     checkdim_(*this, rhs);
     
-    for (uint i = 0u; i < nrow(); ++i) {
-        for (uint j = 0u; j < ncol(); ++j) {
+    for (size_t i = 0u; i < nrow(); ++i) {
+        for (size_t j = 0u; j < ncol(); ++j) {
             this->operator()(i, j) -= rhs.get_cell(i, j);
         }
     }
@@ -78,8 +78,8 @@ BDENSE_TEMPLATE(BDENSE_TYPE()&, operator-=) (
     const Cell_Type& rhs
 ) {
 
-    for (uint i = 0u; i < nrow(); ++i) 
-        for (uint j = 0u; j < ncol(); ++j) 
+    for (size_t i = 0u; i < nrow(); ++i) 
+        for (size_t j = 0u; j < ncol(); ++j) 
             this->operator()(i, j) -= rhs;
         
     
@@ -91,8 +91,8 @@ BDENSE_TEMPLATE(BDENSE_TYPE()&, operator*=) (
     const Cell_Type& rhs
 ) {
 
-    for (uint i = 0u; i < nrow(); ++i) 
-        for (uint j = 0u; j < nrow(); ++j)
+    for (size_t i = 0u; i < nrow(); ++i) 
+        for (size_t j = 0u; j < nrow(); ++j)
             el[POS(i, j)] *= rhs;
 
     return *this;
@@ -102,8 +102,8 @@ BDENSE_TEMPLATE(BDENSE_TYPE()&, operator/=) (
     const Cell_Type& rhs
 ) {
 
-    for (uint i = 0u; i < nrow(); ++i) 
-        for (uint j = 0u; j < nrow(); ++j)
+    for (size_t i = 0u; i < nrow(); ++i) 
+        for (size_t j = 0u; j < nrow(); ++j)
             el[POS(i, j)] /= rhs;
 
     return *this;
